PLAY-216- converted date stacked kit to typescript
