+++ playbook/app/pb_kits/playbook/pb_date_stacked/docs/_date_stacked_default.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import DateStacked from '../_date_stacked.jsx'
+import DateStacked from '../_date_stacked'
 
 const DateStackedDefault = (props) => {
   return (
